refine command line option naming and docs
